优化并增加工具类功能

+++ by122006library/src/main/java/com/by122006library/Utils/MathUtils.java
@@ -86,7 +86,7 @@ public class MathUtils  {
      * @param per
      * @return
      */
-    public static boolean randomHappened(int per){
+    public static boolean randomHappened(float per){
         return Math.random()*100<per;
     }
